Build swiftshader with BOARD_VNDK_VERSION=current
This is using cutils/log.h which is deprecated in O (in favor of
log/log.h) and this is causing build error when building with
BOARD_VNDK_VERSION=current set. However, since swiftshader should be
able to be built with older versions Android some of which don't have
log/log.h, we can't simply change cutils/log.h to log/log.h.

Instead, liblog_headers is added to the header lib dependency (only for
O and beyond) so that log/log.h can be correctly included via
cutils/log.h

Bug: 63135587
Test: choosecombo into sdk_gphone_x86, mma -j inside this dir
Change-Id: I763250e7410025b1dcd7ae210693e3d6bffdb6f1