Migrate to using <log/log.h> since Android O (API 27)
Starting from Android O, <cutils/log.h> is planned to be deprecated in the
future. The use of the original <cutils/log.h> is split into 2 header files.

Per Treble team's (b/78370064) description, <log/log.h> is for native shared
libs or executables, and <android/log.h> is for app JNI.

So a conditional preprocessor is put in place to look at the
ANDROID_PLATFORM_SDK_VERSION definition. If the value is less than 27, means
it's Android N or older, and it ought to include <cutils/log.h>.

Otherwise, it should be 27 (Android O) or any greater version, and it's
supposed to include the new <log/log.h> file.

If ANDROID_PLATFORM_SDK_VERSION is not definied, Andorid build system catches
it and stops the building process with an error message.

Bug: b/116855807
Test: aosp_arm-eng target on API-28, 27, and 26 branches.
Change-Id: I5af4fea62ebf73118fb6a92e4baff18b3f4cfc81